backend/requirements.txt
django==4.0.4
djangorestframework
pyyaml
requests
django-cors-headers
coverage==5.2.1
pillow